requirements.txt
Pillow>=5.4.0
exifread>=2.3.2
tkinterdnd2>=0.4.3
